move under "course" class, add fixtures
