+++ tests/test_course.py
@@ -21,6 +21,7 @@ from canvasapi.enrollment import Enrollment
 from canvasapi.exceptions import RequiredFieldMissing, ResourceDoesNotExist
 from canvasapi.external_feed import ExternalFeed
 from canvasapi.external_tool import ExternalTool
+from canvasapi.lti_resource_link import LTIResourceLink
 from canvasapi.feature import Feature, FeatureFlag
 from canvasapi.file import File
 from canvasapi.folder import Folder
@@ -1889,7 +1890,36 @@ class TestCourse(unittest.TestCase):
         self.assertEqual(len(root_folder_list), 1)
         self.assertIsInstance(root_folder_list[0], Folder)
         self.assertEqual("course_files", root_folder_list[0].name)
+    
+    # create_lti_resource_link()
+    def test_create_lti_resource_link(self, m):
+        register_uris({"lti_resource_link": ["create_lti_resource_link"]}, m)
+        evnt = self.course.create_lti_resource_link(
+           url="https://example.com/lti/launch/content_item/123", title="Test LTI Resource Link", 
+        )
+        self.assertIsInstance(evnt, LTIResourceLink)
+        self.assertEqual(evnt.title, "Test LTI Resource Link")
+        self.assertEqual(evnt.url, "https://example.com/lti/launch/content_item/123")
 
+    # get_lti_resource_links()
+    def test_get_lti_resource_links(self, m):
+        register_uris({"lti_resource_link": ["list_lti_resource_links"]}, m)
+
+        lti_resource_links = self.course.get_lti_resource_links()
+        lti_resource_link_list = [link for link in lti_resource_links]
+        self.assertEqual(len(lti_resource_link_list), 3)
+        self.assertIsInstance(lti_resource_link_list[0], LTIResourceLink)
+
+    # get_lti_resource_link()
+    def test_get_lti_resource_link(self, m):
+        register_uris({"lti_resource_link": ["get_lti_resource_link"]}, m)
+
+        lti_resource_link_by_id = self.course.get_lti_resource_link(45)
+        self.assertIsInstance(lti_resource_link_by_id, LTIResourceLink)
+        self.assertEqual(lti_resource_link_by_id.title, "Test LTI Resource Link")
+        lti_resource_link_by_obj = self.course.get_lti_resource_link(lti_resource_link_by_id)
+        self.assertIsInstance(lti_resource_link_by_obj, LTIResourceLink)
+        self.assertEqual(lti_resource_link_by_obj.title, "Test LTI Resource Link")
 
 @requests_mock.Mocker()
 class TestCourseNickname(unittest.TestCase):
